services/gminiservice/requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
TTS==0.22.0
